+++ pkg/analyzer/CHANGELOG.md
@@ -10,6 +10,8 @@
   use `CompilationUnitElement.libraryImports` instead.
 * Deprecated `LibraryElement.prefixes`,
   use `CompilationUnitElement.libraryImportPrefixes` instead.
+* Deprecated `LibraryElement.parts`,
+  use `CompilationUnitElement.parts` instead.
 
 ## 6.9.0
 * `NormalFormalParameter` now implements `AnnotatedNode`.